==> Project_4_Resources/1_col_data_cleaning.ipynb <==
import pandas as pd

# Read Cost of Living CSV File and Begin Cleaning Data

# Read csv file
col_data_df = pd.read_csv('Resources/Cost_of_Living_Docs/cost_of_living.csv')
col_data_df.head()

# Create a copy of the original DataFrame
col_data_df_copy = col_data_df.copy()

# Drop the 'county_fips', 'case_id', 'housing', 'food_monthly', 'transportation_monthly', 'healthcare_monthly', 'misc_monthly', 'childcare_monthly', 'taxes_monthly', 'total_monthly' columns from copy
col_data_df_copy = col_data_df_copy.drop(columns=['county_fips', 'case_id', 'housing', 'food_monthly', 'transportation_monthly', 'healthcare_monthly', 'misc_monthly', 'childcare_monthly', 'taxes_monthly', 'total_monthly'])

# Verify the column has been dropped
col_data_df_copy.head()

col_data_df_copy.head(10)

# Save the updated DataFrame to a new CSV file
col_data_df_copy.to_csv('Resources/Cost_of_Living_Docs/updated_col_data.csv', index=False)
