==> Project_4_Resources/2_salary_data_cleaning.ipynb <==
import pandas as pd

# Read Average Salary per Occupation CSV File and Begin Cleaning Data

# Read csv file
salary_data_df = pd.read_csv('Resources/OEWS_Docs/salary_data.csv')
salary_data_df.head()

# Create a copy of the original DataFrame
salary_data_df_copy = salary_data_df.copy()

salary_data_df_copy.columns

# Drop the  columns from copy
salary_data_df_copy = salary_data_df_copy.drop(columns=['area', 'area_type', 'naics', 'naics_title', 'i_group', 'own_code', 'o_group'])

# Verify the column has been dropped
salary_data_df_copy.head()

# Apply the string split method to remove everything after the comma
salary_data_df_copy['county'] = salary_data_df_copy['county'].str.split(',').str[0]

salary_data_df_copy.head()

# Save the updated DataFrame to a new CSV file
salary_data_df_copy.to_csv('Resources/OEWS_Docs/updated_salary_data.csv', index=False)
