==> Project_4_Resources/5_pop_pct_diff_county.ipynb <==
import pandas as pd

county_pct_diff_df = pd.read_csv('Resources/Percent_Change_Data/pop_pct_diff_county.csv')

county_pct_diff_df.head()

# Drop the original header and set row 0 as the new header
county_pct_diff_df.columns = county_pct_diff_df.iloc[0]  
county_pct_diff_df = county_pct_diff_df.drop(0)  

# Remove the leading '.' in the 'Geographic Area' column
county_pct_diff_df['Geographic Area'] = county_pct_diff_df['Geographic Area'].str.replace(r'^\.', '', regex=True)

# Split 'Geographic Area' into 'county' and 'state_abv'
county_pct_diff_df[['county', 'state_abv']] = county_pct_diff_df['Geographic Area'].str.split(', ', expand=True)

# Optionally, drop the original 'Geographic Area' column if it's no longer needed
county_pct_diff_df = county_pct_diff_df.drop(columns=['Geographic Area'])

county_pct_diff_df.head()

# Reorder the columns
county_pct_diff_df = county_pct_diff_df[['state_abv', 'county', '2022', '2023', 'pct_diff']]

# Drop '2022' and '2023' columns
county_pct_diff_df = county_pct_diff_df.drop(columns=['2022', '2023'])

county_pct_diff_df.head()

# Define the state-to-abbreviation dictionary
s_state_to_abbrev = {
    "Alabama": "AL",
    "Alaska": "AK",
    "Arizona": "AZ",
    "Arkansas": "AR",
    "California": "CA",
    "Colorado": "CO",
    "Connecticut": "CT",
    "Delaware": "DE",
    "Florida": "FL",
    "Georgia": "GA",
    "Hawaii": "HI",
    "Idaho": "ID",
    "Illinois": "IL",
    "Indiana": "IN",
    "Iowa": "IA",
    "Kansas": "KS",
    "Kentucky": "KY",
    "Louisiana": "LA",
    "Maine": "ME",
    "Maryland": "MD",
    "Massachusetts": "MA",
    "Michigan": "MI",
    "Minnesota": "MN",
    "Mississippi": "MS",
    "Missouri": "MO",
    "Montana": "MT",
    "Nebraska": "NE",
    "Nevada": "NV",
    "New Hampshire": "NH",
    "New Jersey": "NJ",
    "New Mexico": "NM",
    "New York": "NY",
    "North Carolina": "NC",
    "North Dakota": "ND",
    "Ohio": "OH",
    "Oklahoma": "OK",
    "Oregon": "OR",
    "Pennsylvania": "PA",
    "Rhode Island": "RI",
    "South Carolina": "SC",
    "South Dakota": "SD",
    "Tennessee": "TN",
    "Texas": "TX",
    "Utah": "UT",
    "Vermont": "VT",
    "Virginia": "VA",
    "Washington": "WA",
    "West Virginia": "WV",
    "Wisconsin": "WI",
    "Wyoming": "WY",
    "District of Columbia": "DC"
}

# Replace the full state names in 'state_abv' column with their abbreviations
county_pct_diff_df['state_abv'] = county_pct_diff_df['state_abv'].map(s_state_to_abbrev)

county_pct_diff_df.head()

county_pct_diff_df.to_csv('Resources/Percent_Change_Data/county_pct_diff.csv', index=True)
