==> Project_4_Resources/6_create_pct_diff_binary.ipynb <==
import pandas as pd

# Read CSV file 'updated_col_data.csv'
col_df = pd.read_csv('Resources/Cost_of_Living_Docs/updated_col_data.csv')
col_df.head()

# Drop columns 'num_counties_in_st', 'st_cost_rank', 'st_med_aff_rank', 'st_income_rank'
mf_df = col_df.drop(columns=['family', 'num_counties_in_st', 'st_cost_rank', 'st_med_aff_rank', 'st_income_rank'])

# Drop rows with any NaN values
col_df.dropna(inplace=True)

col_df = pd.DataFrame(col_df)

# Remove dollar signs and commas, and convert relevant columns to numeric
columns_to_convert = ['housing_yearly', 'food_yearly', 'transportation_yearly', 
                      'healthcare_yearly', 'misc_yearly', 'childcare_yearly', 
                      'taxes_yearly', 'total_yearly', 'median_family_income']

for column in columns_to_convert:
    # Replace dollar signs and commas
    col_df[column] = col_df[column].replace({'\$': '', ',': ''}, regex=True)

    # Convert to numeric, coercing errors to NaN
    col_df[column] = pd.to_numeric(col_df[column], errors='coerce')

# Check data types of the DataFrame
print(col_df.dtypes)

# Drop rows with NaN values in specific columns
col_df.dropna(subset=columns_to_convert, inplace=True)

# Ensure that the columns are numeric after dropping NaNs
for column in columns_to_convert:
    assert pd.api.types.is_numeric_dtype(col_df[column]), f"{column} is not numeric"

# Group by state_abv and county, then calculate mean
average_df = col_df.groupby(['state_abv', 'county'])[columns_to_convert].mean().reset_index()

average_df.head()

pct_diff_df = pd.read_csv('Resources/Percent_Change_Data/county_pct_diff.csv')
pct_diff_df.head()

merge_df = pd.merge(average_df, pct_diff_df, on=['state_abv', 'county'], how='inner')

merge_df.head()

# Find mean for 'pct_diff' column
mean_pct_diff = pct_diff_df['pct_diff'].mean()

print(f"Mean of 'pct_diff' column: {mean_pct_diff}")

# Step 1: Convert pct_diff to binary values in merge_df
merge_df['pct_diff_binary'] = merge_df['pct_diff'].apply(lambda x: 0 if x > mean_pct_diff else 1)

merge_df.head()

merge_df.to_csv('Resources/Percent_Change_Data/binary_county_pct_diff.csv', index=True)
